+++ hw03/src/utils.cuh
@@ -5,7 +5,7 @@
 #include <cuda_device_runtime_api.h>
 #include <driver_types.h>
 
-void initialize_cu(double *marr, double *xarr, int N, int dim, int Tx, int Ty);
-void compute_cu(double *xarr, int nsteps, int N, int dim, double G, double dt, double cut);
+void initialize_cu(float *marr, float *xarr, int N, int dim, int Tx, int Ty);
+void compute_cu(float *xarr, int nsteps, int N, int dim, float G, float dt, float cut);
 void finalize_cu();
